Move server info to theming section
