+++ tests/lib/Settings/ManagerTest.php
@@ -72,13 +72,14 @@ class ManagerTest extends TestCase {
 
 		$this->manager->registerSection('admin', \OCA\WorkflowEngine\Settings\Section::class);
 
-		$this->url->expects($this->exactly(6))
+		$this->url->expects($this->exactly(7))
 			->method('imagePath')
 			->willReturnMap([
 				['settings', 'admin.svg', '0'],
 				['core', 'actions/settings-dark.svg', '1'],
 				['core', 'actions/share.svg', '2'],
 				['core', 'actions/password.svg', '3'],
+				['settings', 'theming-dark.svg', '6'],
 				['core', 'places/contacts.svg', '5'],
 				['settings', 'help.svg', '4'],
 			]);
@@ -88,6 +89,7 @@ class ManagerTest extends TestCase {
 			1 => [new Section('server', 'Basic settings', 0, '1')],
 			5 => [new Section('sharing', 'Sharing', 0, '2')],
 			10 => [new Section('security', 'Security', 0, '3')],
+			30 => [new Section('theming', 'Theming', 0, '6')],
 			50 => [new Section('groupware', 'Groupware', 0, '5')],
 			55 => [\OC::$server->query(\OCA\WorkflowEngine\Settings\Section::class)],
 			98 => [new Section('additional', 'Additional settings', 0, '1')],
@@ -124,13 +126,14 @@ class ManagerTest extends TestCase {
 			->method('t')
 			->will($this->returnArgument(0));
 
-		$this->url->expects($this->exactly(6))
+		$this->url->expects($this->exactly(7))
 			->method('imagePath')
 			->willReturnMap([
 				['settings', 'admin.svg', '0'],
 				['core', 'actions/settings-dark.svg', '1'],
 				['core', 'actions/share.svg', '2'],
 				['core', 'actions/password.svg', '3'],
+				['settings', 'theming-dark.svg', '6'],
 				['core', 'places/contacts.svg', '5'],
 				['settings', 'help.svg', '4'],
 			]);
@@ -140,6 +143,7 @@ class ManagerTest extends TestCase {
 			1 => [new Section('server', 'Basic settings', 0, '1')],
 			5 => [new Section('sharing', 'Sharing', 0, '2')],
 			10 => [new Section('security', 'Security', 0, '3')],
+			30 => [new Section('theming', 'Theming', 0, '6')],
 			50 => [new Section('groupware', 'Groupware', 0, '5')],
 			98 => [new Section('additional', 'Additional settings', 0, '1')],
 		], $this->manager->getAdminSections());
@@ -209,7 +213,7 @@ class ManagerTest extends TestCase {
 		$this->manager->registerSection('personal', \OCA\WorkflowEngine\Settings\Section::class);
 		$this->manager->registerSection('admin', \OCA\WorkflowEngine\Settings\Section::class);
 
-		$this->url->expects($this->exactly(9))
+		$this->url->expects($this->exactly(10))
 			->method('imagePath')
 			->willReturnMap([
 				['core', 'actions/info.svg', '1'],
@@ -219,6 +223,7 @@ class ManagerTest extends TestCase {
 				['core', 'actions/settings-dark.svg', '1'],
 				['core', 'actions/share.svg', '2'],
 				['core', 'actions/password.svg', '3'],
+				['settings', 'theming-dark.svg', '6'],
 				['core', 'places/contacts.svg', '5'],
 				['settings', 'help.svg', '4'],
 			]);
@@ -235,6 +240,7 @@ class ManagerTest extends TestCase {
 			1 => [new Section('server', 'Basic settings', 0, '1')],
 			5 => [new Section('sharing', 'Sharing', 0, '2')],
 			10 => [new Section('security', 'Security', 0, '3')],
+			30 => [new Section('theming', 'Theming', 0, '6')],
 			50 => [new Section('groupware', 'Groupware', 0, '5')],
 			55 => [\OC::$server->query(\OCA\WorkflowEngine\Settings\Section::class)],
 			98 => [new Section('additional', 'Additional settings', 0, '1')],
